Server class
